==> scripts/2_CGL-NGRAMS.ipynb <==
import numpy as np
import pandas as pd
import requests
import re
import nltk
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import TruncatedSVD
from sklearn.metrics.pairwise import cosine_similarity
import pickle
from cltk.stops.grc import STOPS as stopwords
from gensim.corpora import Dictionary
import unicodedata
import json

# to communicate with google spreadsheet...
import gspread
from gspread_dataframe import get_as_dataframe
from gspread_dataframe import set_with_dataframe
from google.oauth2 import service_account # based on google-auth library

file_data = json.load(open("../../ServiceAccountsKey.json", "r"))
# or load it from a local storage: json.load(open("../../ServiceAccountsKey.json", "r"))
credentials = service_account.Credentials.from_service_account_info(file_data)
gc = gspread.Client(auth=credentials.with_scopes(['https://spreadsheets.google.com/feeds', 'https://www.googleapis.com/auth/drive']))
PIPA_data = gc.open_by_url("https://docs.google.com/spreadsheets/d/1rV4t0_UV_wcx--UAHVwkqB8Wa_5n9mnpV05yGG1OHqk/edit?usp=sharing")

# Load the main dataset of ancient Greek texts

cgl = pd.read_json("../data/large_data/cgl.json")
cgl.head()

cgl["lemmata"].apply(lambda x: x.count("συζάω")).sum()

sents = [sen for work in cgl["lemmatized_sentences_repl"] for sen in work]

len(sents)

min_freq = 50

def get_vocab(docs, min_freq=min_freq):
    words_flat = [item for sublist in docs for item in sublist]
    word_freq_tups = nltk.FreqDist(words_flat).most_common()
    vocabulary = [tup[0] for tup in word_freq_tups if tup[1] >= min_freq]
    #vocab_freqs = [len([doc for doc in docs if word in doc]) for word in vocabulary]
    return word_freq_tups, words_flat, vocabulary # , vocab_freqs

word_freqs, words, vocabulary = get_vocab(sents)

word_freqs[:10]

len(words)

len(vocabulary)

print(vocabulary[:20])

def normalize_encoding(string):
    return unicodedata.normalize("NFC", string)

stopwords = [normalize_encoding(w) for w in stopwords] + ["εἰμί", "οὗτος", "αὐτός", "αὐτός", "τις", "τίς"]

vocabulary = [w for w in vocabulary if w not in stopwords]
vocabulary[:10]

len(vocabulary)

# continuous bigrams & trigrams, within sentences only
sents_bigrams = [list(el) for sublist in [[ng for ng in nltk.bigrams(sent)] for sent in sents] for el in sublist]

sents_trigrams = [list(el) for sublist in [[ng for ng in nltk.trigrams(sent)] for sent in sents] for el in sublist]

ngrams_data = sents + sents_bigrams + sents_trigrams
len(ngrams_data)

author_ids = list(set(cgl["author_id"]))
author_ids[:5]

with open("../data/large_data/corpus_ngrams.txt", "w", encoding="utf-8") as f:

    f.writelines("\n".join([" ".join(ngram) for ngram in ngrams_data]))

pickle.dump(vocabulary, open("../data/vocabulary.pickle", "wb"))

pickle.dump(word_freqs, open("../data/word_freqs_all.pickle", "wb"))

word_freqs_vocabulary = pd.DataFrame([tup for tup in word_freqs if tup[0] in vocabulary], columns=["word", "count"])

word_freqs_vocabulary.to_csv("../data/word_freqs_vocabulary.csv", index=False)

# Ngrams by doc ids

doc_ids = list(set(cgl["doc_id"]))
len(doc_ids)

cgl_subset = cgl[cgl["doc_id"]=="tlg0086.tlg010"]
sents = [[sen for work in cgl_subset["lemmatized_sentences_repl"] for sen in work][0]]
sents_bigrams = [list(el) for sublist in [[ng for ng in nltk.bigrams(sent)] for sent in sents] for el in sublist]
sents_trigrams = [list(el) for sublist in [[ng for ng in nltk.trigrams(sent)] for sent in sents] for el in sublist]
ngrams_data = sents + sents_bigrams + sents_trigrams

" ".join(sents[0])

[" ".join(bigram) for bigram in sents_bigrams]

sents_trigrams

cgl_subset["clean_string"].tolist()[0][:200]

line = 0
ids_lines = {}

f = open("../data/large_data/corpus_ngrams_bydocid.txt", "w", encoding="utf-8")
for doc_id in doc_ids:
    cgl_subset = cgl[cgl["doc_id"]==doc_id]
    sents = [sen for work in cgl_subset["lemmatized_sentences_repl"] for sen in work]
    sents_bigrams = [list(el) for sublist in [[ng for ng in nltk.bigrams(sent)] for sent in sents] for el in sublist]
    sents_trigrams = [list(el) for sublist in [[ng for ng in nltk.trigrams(sent)] for sent in sents] for el in sublist]
    ngrams_data = sents + sents_bigrams + sents_trigrams
    ngrams_data = [" ".join(ngram) for ngram in ngrams_data]
    f.writelines("\n".join(ngrams_data)+"\n")
    if bool(ngrams_data):
        ids_lines[doc_id] = (line, line+len(ngrams_data))
        line += len(ngrams_data)
    else:
        ids_lines[doc_id] = (line, line+1)
        line += 1

ids_lines

pickle.dump(ids_lines, open("../data/ids_lines.pickle", "wb"))

# Testing that the indexing works

with open("../data/large_data/corpus_ngrams_bydocid.txt", "r") as f:
    lines = f.readlines()
    doc_vals = ids_lines["tlg0086.tlg014"]
    for line in lines[doc_vals[0]:doc_vals[1]][:10]:
        if bool(line.split()):
            print(line.split())

cgl[cgl["doc_id"]=="tlg0086.tlg014"]["lemmatized_sentences"]

c_aristotelicum_ids = cgl[cgl["subcorpus"]=="c_aristotelicum"]["doc_id"].tolist()
c_aristotelicum_ids[:5]

ids_list = c_aristotelicum_ids[:5]

class NgramCorpus:
    def __iter__(self):
        with open("../data/large_data/corpus_ngrams_bydocid.txt", "r") as f:
            lines = f.readlines()
            for phi_id in ids_list:
                for line in lines[ids_lines[phi_id][0] : ids_lines[phi_id][1]]:
                    if bool(line.split()):
                        yield line.split()

corpus = NgramCorpus()
[doc for doc in corpus][:5]

dct = Dictionary([vocabulary])

test_doc = [doc for doc in corpus][0]
print(test_doc)

dct.doc2bow(test_doc)

class NgramCorpus(object):
    def __init__(self, ids_list, ids_lines, fname, bow=False, dct=None):
        self.ids_list = ids_list
        self.fname = fname
        self.bow  = bow
        self.dct = dct
        self.ids_lines = ids_lines
        self.len = 0
    def __len__(self):
        [el for el in self]
        return self.len
    def __iter__(self):
        with open(self.fname, "r") as f:
            lines = f.readlines()
            for id in self.ids_list:
                for line in lines[self.ids_lines[id][0] : self.ids_lines[id][1]]:
                    if bool(line.split()):
                        self.len += 1
                        if (self.bow) & (self.dct != None):
                            yield self.dct.doc2bow(line.split())
                        else:
                            yield line.split()

corpus = NgramCorpus(c_aristotelicum_ids[:4], ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt", bow=True, dct=dct)

len(corpus)

[doc for doc in corpus][0]
