==> scripts/3_GENERATING-VECTORS.ipynb <==
from gensim.corpora import Dictionary
from gensim.models import Word2Vec
import pandas as pd
import pickle

cgl = pd.read_json("../data/large_data/cgl.json")

word_freqs_vocabulary = pd.read_csv("../data/word_freqs_vocabulary.csv")

word_freqs_vocabulary

word_freq_dict = dict(zip(word_freqs_vocabulary["word"], word_freqs_vocabulary["count"]))

len(word_freq_dict)

vocabulary = word_freqs_vocabulary["word"].tolist()

ids_lines = pickle.load(open("../data/ids_lines.pickle", "rb"))

class NgramCorpus(object):
    def __init__(self, ids_list, ids_lines, fname, bow=False, dct=None):
        self.ids_list = ids_list
        self.fname = fname
        self.bow  = bow
        self.dct = dct
        self.ids_lines = ids_lines
        self.len = 0
    def __len__(self):
        [el for el in self]
        return self.len
    def __iter__(self):
        with open(self.fname, "r") as f:
            lines = f.readlines()
            for id in self.ids_list:
                for line in lines[self.ids_lines[id][0] : self.ids_lines[id][1]]:
                    if bool(line.split()):
                        self.len += 1
                        if (self.bow) & (self.dct != None):
                            yield self.dct.doc2bow(line.split())
                        else:
                            yield line.split()

# Full model

corpus_full = NgramCorpus(cgl["doc_id"].tolist(), ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl["lemmata_wordcount"].sum()

len(corpus_full)

def from_corpus_to_keyed_vectors(corpus):
    model = Word2Vec(vector_size=150, window=3, negative=5, ns_exponent=1, sg=0, epochs=5, workers=8)
    model.build_vocab_from_freq(word_freq=word_freq_dict)
    model.train(corpus, total_examples=len(corpus), epochs=model.epochs)
    keyed_vectors = model.wv
    return keyed_vectors

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_full)

keyed_vectors.save("../data/keyed_vectors_full.wv")

# Model based only on subcorpora

# ids_list = cgl[cgl["subcorpus"] != "c_aristotelicum"]["doc_id"].tolist()
ids_list = cgl[cgl["author_id"].isin(["tlg0086", "tlg0627", "tlg0059"])]["doc_id"].tolist()
print(len(ids_list))
corpus_subcorpora = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl[cgl["author_id"].isin(["tlg0086", "tlg0627", "tlg0059"])]["lemmata_wordcount"].sum()

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_subcorpora)
keyed_vectors.save("../data/keyed_vectors_subcorpora.wv")

# Model excluding all subcorpora

# ids_list = cgl[cgl["subcorpus"] != "c_aristotelicum"]["doc_id"].tolist()
ids_list = cgl[~cgl["author_id"].isin(["tlg0086", "tlg0627", "tlg0059"])]["doc_id"].tolist()
print(len(ids_list))
corpus_excl_subcorpora = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_excl_subcorpora)
keyed_vectors.save("../data/keyed_vectors_excl_subcorpora.wv")

# Model exclusing c_aristotelicum & c_platonicum

# ids_list = cgl[cgl["subcorpus"] != "c_aristotelicum"]["doc_id"].tolist()
ids_list = cgl[~cgl["author_id"].isin(["tlg0086", "tlg0059"])]["doc_id"].tolist()
print(len(ids_list))
corpus_excl_aristplato = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl[~cgl["author_id"].isin(["tlg0086", "tlg0059"])]["lemmata_wordcount"].sum()

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_excl_aristplato)
keyed_vectors.save("../data/keyed_vectors_excl_aristplato.wv")

# Model excluding c_aristotelicum

set(cgl["subcorpus"])

# ids_list = cgl[cgl["subcorpus"] != "c_aristotelicum"]["doc_id"].tolist()
ids_list = cgl[cgl["author_id"] != "tlg0086"]["doc_id"].tolist()
print(len(ids_list))
corpus_excl_arist = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl[cgl["author_id"] != "tlg0086"]["lemmata_wordcount"].sum()

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_excl_arist)
keyed_vectors.save("../data/keyed_vectors_excl_arist.wv")

# Model excluding c_hippocraticum

# ids_list = cgl[cgl["subcorpus"] != "c_hippocraticum"]["doc_id"].tolist()
ids_list = cgl[cgl["author_id"] != "tlg0627"]["doc_id"].tolist()
print(len(ids_list))
corpus_excl_hipp = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl[cgl["author_id"] != "tlg0627"]["lemmata_wordcount"].sum()

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_excl_hipp)
keyed_vectors.save("../data/keyed_vectors_excl_hipp.wv")

# Model excluding c_platonicum

# ids_list = cgl[cgl["subcorpus"] != "c_platonicum"]["doc_id"].tolist()
ids_list = cgl[cgl["author_id"] != "tlg0059"]["doc_id"].tolist()
print(len(ids_list))
corpus_excl_plato = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

cgl[cgl["author_id"] != "tlg0059"]["lemmata_wordcount"].sum()

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_excl_plato)
keyed_vectors.save("../data/keyed_vectors_excl_plato.wv")

# Aristotle alone...

# ids_list = cgl[cgl["subcorpus"] != "c_aristotelicum"]["doc_id"].tolist()
ids_list = cgl[cgl["author_id"] == "tlg0086"]["doc_id"].tolist()
print(len(ids_list))
corpus_arist = NgramCorpus(ids_list, ids_lines, "../data/large_data/corpus_ngrams_bydocid.txt")

%%time
keyed_vectors = from_corpus_to_keyed_vectors(corpus_arist)
keyed_vectors.save("../data/keyed_vectors_arist.wv")

cgl[cgl["author_id"] == "tlg0086"]["wordcount"].sum()

cgl[cgl["author_id"] == "tlg0086"]["lemmata_wordcount"].sum()

len(cgl)

cgl["wordcount"].sum()

cgl["lemmata_wordcount"].sum()
