==> scripts/5_CATEGORIES-DEVELOPMENT.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

from gensim.models import KeyedVectors
from sklearn.metrics.pairwise import cosine_similarity

# plotting
import matplotlib.pyplot as plt
import matplotlib.patches as patches
import matplotlib.colors as mcolors

import seaborn as sns
import os
import json
from translator import translator
import pickle

# to communicate with google spreadsheet...
import gspread
from gspread_dataframe import get_as_dataframe
from gspread_dataframe import set_with_dataframe
from google.oauth2 import service_account # based on google-auth library

# establish connection with gogglesheets...
file_data = json.load(open("../../ServiceAccountsKey.json", "r"))
credentials = service_account.Credentials.from_service_account_info(file_data)
gc = gspread.Client(auth=credentials.with_scopes(['https://spreadsheets.google.com/feeds', 'https://www.googleapis.com/auth/drive']))
PIPA_data = gc.open_by_url("https://docs.google.com/spreadsheets/d/1rV4t0_UV_wcx--UAHVwkqB8Wa_5n9mnpV05yGG1OHqk/edit?usp=sharing")

keywords = ["ἀλγέω", "ἄλγημα", "ἄλγος", "λυπέω", "λύπη", "λυπηρός", "ὀδυνάω", "ὀδύνη", "πονέω", "πόνος"]

vectors_dict = {}
for filename in [f for f in os.listdir("../data/") if "keyed_vectors" in f]:
    vectors_dict[filename.partition(".")[0]] = KeyedVectors.load("../data/" + filename)

complete_sim_matrices = {}
for item in vectors_dict.items():
    complete_sim_matrix = cosine_similarity(item[1].vectors)
    complete_sim_matrices[item[0].replace("keyed_vectors_", "")] = complete_sim_matrix

complete_sim_matrices

# Analyzing categories

terms_translation_categories = get_as_dataframe(PIPA_data.worksheet("translation"))[["greek", "english", "category_clean"]]
terms_translation_categories = terms_translation_categories[terms_translation_categories["greek"].notnull()]
terms_translation_categories.rename(columns={"category_clean" : "category"}, inplace=True)
terms_translation_categories.head(5)

terms_translation_categories["greek"] = terms_translation_categories["greek"].apply(lambda x: x.rpartition(" (")[0].replace(" ", ""))

terms_translation_categories = terms_translation_categories.drop_duplicates(subset="greek")

terms_translation_categories.groupby("category").size()

categories = list(set(terms_translation_categories["category"]))

term_category_dict = dict(zip(terms_translation_categories["greek"], terms_translation_categories["category"]))
term_category_dict

term_translation_dict = dict(zip(terms_translation_categories["greek"], terms_translation_categories["english"]))
term_translation_dict

cat_terms_dict = {}
for cat in categories:
    terms = terms_translation_categories[terms_translation_categories["category"]==cat]["greek"].tolist()
    cat_terms_dict[cat] = terms

cat_terms_dict

vectors_dict["keyed_vectors_full"].n_similarity(["πόνος"], cat_terms_dict["dietetics"])

sim_tups = []
for keyword in keywords:
    for cat in cat_terms_dict.keys():
        sim_tups.append((keyword, cat, vectors_dict["keyed_vectors_full"].n_similarity([keyword], cat_terms_dict[cat])))

sim_tups

# Extending categories

other_words = []
for term in term_category_dict.keys():
    other_words.extend([tup[0] for tup in vectors_dict["keyed_vectors_full"].most_similar(term, topn=5)])

len(other_words)

other_words_unique =  set(other_words)
len(other_words_unique)

other_words_unique_new = list(other_words_unique.difference(term_category_dict.keys()))
other_words_unique_new[:10]

word_cat_extension = pd.DataFrame({"greek" : other_words_unique_new, "english" : [translator(word) for word in other_words_unique_new], "category" : ["" for el in other_words_unique_new]})

#set_with_dataframe(PIPA_data.add_worksheet("word_cat_extension", 1,1), word_cat_extension)

# Load back the manually coded extension

word_freqs = pickle.load(open("../data/word_freqs_all.pickle", "rb"))
word_freqs_dict = dict(word_freqs)

word_freqs_dict["ἀγαθός"]

word_cat_extension_coded = get_as_dataframe(PIPA_data.worksheet("word_cat_extension"))[["greek", "english", "category"]]
word_cat_extension_coded = word_cat_extension_coded[word_cat_extension_coded["category"].notnull()]
word_cat_extension_coded.head(5)

len(word_cat_extension_coded)

word_cat_extension2_coded = get_as_dataframe(PIPA_data.worksheet("word_cat_extension2"))[["greek", "english", "category"]]
word_cat_extension2_coded = word_cat_extension2_coded[word_cat_extension2_coded["category"].notnull()]
word_cat_extension2_coded.head(5)

len(word_cat_extension2_coded)

terms_translation_categories = pd.concat([terms_translation_categories, word_cat_extension_coded, word_cat_extension2_coded])

terms_translation_categories = terms_translation_categories[~terms_translation_categories["category"].str.contains("other")]

terms_translation_categories["count"] = terms_translation_categories["greek"].apply(lambda x: word_freqs_dict[x])
terms_translation_categories.sort_values("count", ascending=False, inplace=True)
terms_translation_categories.head(5)

len(terms_translation_categories)

terms_translation_categories.drop_duplicates(subset="greek", inplace=True)
len(terms_translation_categories)

terms_translation_categories["category"].replace("emotion", "emotions", inplace=True)
terms_translation_categories["category"].replace("moral", "morality", inplace=True)

pickle.dump(terms_translation_categories, open("../data/terms_translation_categories.pickle", "wb"))

terms_translation_categories.to_csv("../data/terms_translation_categories.csv")

set_with_dataframe(PIPA_data.add_worksheet("terms_translation_categories", 1,1), terms_translation_categories)
